fixed bug that left behind the json of files associated with and unusable run

+++ oceanproc/bids_wrapper.py
@@ -85,9 +85,11 @@ def remove_unusable_runs(xml_file:Path, bids_path:Path, subject:str, session:str
         if quality_pairs[j_key] == "unusable":
             logger.info(f"  Removing series {j_key[0]} - {j_key[1]} - {j_key[2]}: \n\t NIFTI:{file.path}")
             os.remove(file.path)
-            for assoc_file in file.get_associations():
-                logger.info(f"      removing associated file: {assoc_file.path}")
-                os.remove(assoc_file.path)
+            assoc_list = set(list(file.get_associations()) + [af for assoc in file.get_associations() for af in assoc.get_associations()])
+            for assoc_file in assoc_list:
+                if Path(assoc_file.path).exists():
+                    logger.info(f"      removing associated file: {assoc_file.path}")
+                    os.remove(assoc_file.path)
 
 
 @debug_logging
